Split test_ote_training.py into several files
