+++ tests/ote_training_tests_helper.py
@@ -1,0 +1,204 @@
+import logging
+import itertools
+from abc import ABC, abstractmethod
+from copy import deepcopy
+from collections import OrderedDict
+from typing import Any, Dict, List, Optional, Type
+
+from ote_training_test_case import (OTETestCaseInterface,
+                                    generate_ote_integration_test_case_class,
+                                    )
+from ote_training_tests_common import DEFAULT_FIELD_VALUE_FOR_USING_IN_TEST
+from ote_training_tests_actions import get_default_test_action_classes
+
+
+logger = logging.getLogger(__name__)
+
+
+class OTETrainingTestInterface(ABC):
+    @classmethod
+    @abstractmethod
+    def get_list_of_tests(cls, usecase: Optional[str] = None):
+        raise NotImplementedError('The method is not implemented')
+
+class OTETestCreationParametersInterface(ABC):
+    @abstractmethod
+    def test_bunches(self) -> List[Dict[str, Any]]:
+        raise NotImplementedError('The method is not implemented')
+
+    @abstractmethod
+    def test_case_class(self) -> Type[OTETestCaseInterface]:
+        raise NotImplementedError('The method is not implemented')
+
+    @abstractmethod
+    def short_test_parameters_names_for_generating_id(self) -> OrderedDict:
+        raise NotImplementedError('The method is not implemented')
+
+    @abstractmethod
+    def test_parameters_defining_impl_behavior(self) -> List[str]:
+        raise NotImplementedError('The method is not implemented')
+
+    @abstractmethod
+    def default_test_parameters(self) -> Dict[str, Any]:
+        raise NotImplementedError('The method is not implemented')
+
+class DefaultOTETestCreationParametersInterface(OTETestCreationParametersInterface):
+    def test_case_class(self) -> Type[OTETestCaseInterface]:
+        return generate_ote_integration_test_case_class(get_default_test_action_classes())
+
+    def short_test_parameters_names_for_generating_id(self) -> OrderedDict:
+        DEFAULT_SHORT_TEST_PARAMETERS_NAMES_FOR_GENERATING_ID = OrderedDict([
+                ('test_stage', 'ACTION'),
+                ('model_name', 'model'),
+                ('dataset_name', 'dataset'),
+                ('num_training_iters', 'num_iters'),
+                ('batch_size', 'batch'),
+                ('usecase', 'usecase'),
+        ])
+        return deepcopy(DEFAULT_SHORT_TEST_PARAMETERS_NAMES_FOR_GENERATING_ID)
+
+    def test_parameters_defining_impl_behavior(self) -> List[str]:
+        DEFAULT_TEST_PARAMETERS_DEFINING_IMPL_BEHAVIOR = ['model_name',
+                                                          'dataset_name',
+                                                          'num_training_iters',
+                                                          'batch_size']
+        return deepcopy(DEFAULT_TEST_PARAMETERS_DEFINING_IMPL_BEHAVIOR)
+
+    def default_test_parameters(self) -> Dict[str, Any]:
+        DEFAULT_TEST_PARAMETERS = {
+                'num_iters': 1,
+                'batch_size': 2,
+        }
+        return deepcopy(DEFAULT_TEST_PARAMETERS)
+
+class OTETestHelper:
+    class _Cache:
+        def __init__(self):
+            self._cache_parameters = {}
+            self._cached_value = None
+        def get(self):
+            return self._cached_value
+        def set(self, params, value):
+            logger.debug(f'cache.set new value for parameters {params}')
+            self._cache_parameters = deepcopy(params)
+            self._cached_value = value
+        def has_same_params(self, params):
+            res = (self._cache_parameters == params)
+            res_str = '==' if res else '!='
+            logger.debug(f'cache.has_same_params: '
+                         f'cache_parameters={self._cache_parameters} {res_str} {params}, res={res}')
+            return res
+
+    def __init__(self,
+                 test_creation_parameters):
+        assert isinstance(test_creation_parameters, OTETestCreationParametersInterface)
+
+        self.test_case_class = test_creation_parameters.test_case_class()
+        self.test_bunches = test_creation_parameters.test_bunches()
+
+        self.short_test_parameters_names_for_generating_id = \
+                test_creation_parameters.short_test_parameters_names_for_generating_id()
+        #TODO(beynens): rename to test_parameters_defining_test_case
+        self.test_parameters_defining_impl_behavior = \
+                test_creation_parameters.test_parameters_defining_impl_behavior()
+        self.default_test_parameters = \
+                test_creation_parameters.default_test_parameters()
+
+        self._cache = OTETestHelper._Cache()
+
+        assert issubclass(self.test_case_class, OTETestCaseInterface)
+        assert isinstance(self.short_test_parameters_names_for_generating_id, OrderedDict)
+        assert all(isinstance(k, str) and isinstance(v, str)
+                   for k, v in self.short_test_parameters_names_for_generating_id.items())
+        assert 'test_stage' in self.short_test_parameters_names_for_generating_id
+        assert 'model_name' in self.short_test_parameters_names_for_generating_id
+        assert 'dataset_name' in self.short_test_parameters_names_for_generating_id
+        assert 'usecase' in self.short_test_parameters_names_for_generating_id
+
+        assert isinstance(self.test_parameters_defining_impl_behavior, list)
+        assert all(isinstance(s, str) for s in self.test_parameters_defining_impl_behavior)
+
+        assert isinstance(self.default_test_parameters, dict)
+        assert all(isinstance(k, str) for k in self.default_test_parameters.keys())
+
+    def _get_list_of_test_stages(self):
+        return self.test_case_class.get_list_of_test_stages()
+
+    def _fill_test_parameters_default_values(self, test_parameters):
+        for key, default_val in self.default_test_parameters.items():
+            val = test_parameters.get(key)
+            if val is None or val == DEFAULT_FIELD_VALUE_FOR_USING_IN_TEST():
+                test_parameters[key] = default_val
+
+    def _generate_test_id(self, test_parameters):
+        id_parts = (
+                f'{short_par_name}-{test_parameters[par_name]}'
+                for par_name, short_par_name in self.short_test_parameters_names_for_generating_id.items()
+        )
+        return ','.join(id_parts)
+
+    def get_list_of_tests(self, usecase: Optional[str] = None):
+        """
+        The functions generates the lists of values for the tests from the field test_bunches of the class.
+
+        The function returns two lists
+        * argnames -- a tuple with names of the test parameters, at the moment it is
+                      a one-element tuple with the parameter name "test_parameters"
+        * argvalues -- list of tuples, each tuple has the same len as argname tuple,
+                       at the moment it is a one-element tuple with the dict `test_parameters`
+                       that stores the parameters of the test
+        * ids -- list of strings with ids corresponding the parameters of the tests
+                 each id is a string generated from the corresponding test_parameters
+                 value -- see the functions _generate_test_id
+
+        The lists argvalues and ids will have the same length.
+
+        If the parameter `usecase` is set, it makes filtering by usecase field of test bunches.
+        """
+        test_bunches = self.test_bunches
+        assert all(isinstance(el, dict) for el in test_bunches)
+
+        argnames = ('test_parameters',)
+        argvalues = []
+        ids = []
+        for el in test_bunches:
+            el_model_name = el.get('model_name')
+            el_dataset_name = el.get('dataset_name')
+            el_usecase = el.get('usecase')
+            if usecase is not None and el_usecase != usecase:
+                continue
+            if isinstance(el_model_name, (list, tuple)):
+                model_names = el_model_name
+            else:
+                model_names = [el_model_name]
+            if isinstance(el_dataset_name, (list, tuple)):
+                dataset_names = el_dataset_name
+            else:
+                dataset_names = [el_dataset_name]
+
+            model_dataset_pairs = list(itertools.product(model_names, dataset_names))
+
+            for m, d in model_dataset_pairs:
+                for test_stage in self._get_list_of_test_stages():
+                    test_parameters = deepcopy(el)
+                    test_parameters['test_stage'] = test_stage
+                    test_parameters['model_name'] = m
+                    test_parameters['dataset_name'] = d
+                    self._fill_test_parameters_default_values(test_parameters)
+                    argvalues.append((test_parameters,))
+                    ids.append(self._generate_test_id(test_parameters))
+
+        return argnames, argvalues, ids
+
+    def get_test_case(self, test_parameters, params_factories_for_test_actions):
+        params_defining_cache = {k: test_parameters[k] for k in self.test_parameters_defining_impl_behavior}
+        if not self._cache.has_same_params(params_defining_cache):
+            logger.info(f'OTETestHelper: parameters were changed -- updating cache')
+            logger.info(f'OTETestHelper: before creating test case (class {self.test_case_class})')
+            test_case = self.test_case_class(params_factories_for_test_actions)
+            logger.info(f'OTETestHelper: after creating test case (class {self.test_case_class})')
+            self._cache.set(params_defining_cache, test_case)
+        else:
+            logger.info('OTETestHelper: parameters were not changed -- cache is kept')
+
+        return self._cache.get()
