+++ tests/test_ote_training.py
@@ -17,11 +17,10 @@ import itertools
 import logging
 import os
 import os.path as osp
-from abc import ABC, abstractmethod
 from collections import Counter, namedtuple, OrderedDict
 from copy import deepcopy
 from pprint import pformat
-from typing import Any, Callable, Dict, List, Optional, Type, Union
+from typing import Any, Callable, Dict, List, Optional, Tuple, Type, Union
 
 import pytest
 import yaml
@@ -30,7 +29,6 @@ from ote_sdk.configuration.helper import create
 from ote_sdk.entities.datasets import DatasetEntity
 from ote_sdk.entities.inference_parameters import InferenceParameters
 from ote_sdk.entities.label_schema import LabelSchemaEntity
-from ote_sdk.entities.metrics import Performance, ScoreMetric
 from ote_sdk.entities.model import (
     ModelEntity,
     ModelFormat,
@@ -51,39 +49,30 @@ from mmdet.apis.ote.apis.detection.ote_utils import get_task_class
 from mmdet.apis.ote.extension.datasets.data_utils import load_dataset_items_coco_format
 from mmdet.integration.nncf.utils import is_nncf_enabled
 
+from ote_training_tests_common import (KEEP_CONFIG_FIELD_VALUE,
+                                      REALLIFE_USECASE_CONSTANT,
+                                      )
+from ote_training_tests_helper import (OTETestHelper,
+                                       DefaultOTETestCreationParametersInterface,
+                                       OTETrainingTestInterface)
+
+
+
+
+from abc import ABC, abstractmethod
+
 logger = logging.getLogger(__name__)
 
 
-def DEFAULT_FIELD_VALUE_FOR_USING_IN_TEST():
-    """
-    This string constant will be used as a special constant for a config field
-    value to point that the field should be filled in tests' code by some default
-    value specific for this field.
-    """
-    return 'DEFAULT_FIELD_VALUE_FOR_USING_IN_TEST'
 
-def KEEP_CONFIG_FIELD_VALUE():
-    """
-    This string constant will be used as a special constant for a config field value to point
-    that the field should NOT be changed in tests -- its value should be taken
-    from the template file or the config file of the model.
-    """
-    return 'KEEP_CONFIG_FIELD_VALUE'
-
-def REALLIFE_USECASE_CONSTANT():
-    """
-    This is a constant for pointing usecase for reallife training tests
-    """
-    return 'reallife'
-
-def DATASET_PARAMETERS_FIELDS():
-    return ('annotations_train',
-            'images_train_dir',
-            'annotations_val',
-            'images_val_dir',
-            'annotations_test',
-            'images_test_dir',
-            )
+def DATASET_PARAMETERS_FIELDS() -> List[str]:
+    return deepcopy(['annotations_train',
+                     'images_train_dir',
+                     'annotations_val',
+                     'images_val_dir',
+                     'annotations_test',
+                     'images_test_dir',
+                     ])
 
 ROOT_PATH_KEY = '_root_path'
 DatasetParameters = namedtuple('DatasetParameters', DATASET_PARAMETERS_FIELDS())
@@ -296,34 +285,6 @@ def _get_dataset_params_from_dataset_definitions(dataset_definitions, dataset_na
     params = DatasetParameters(**training_parameters_fields)
     return params
 
-def performance_to_score_name_value(perf: Union[Performance, None]):
-    """
-    The method is intended to get main score info from Performance class
-    """
-    if perf is None:
-        return None, None
-    assert isinstance(perf, Performance)
-    score = perf.score
-    assert isinstance(score, ScoreMetric)
-    assert isinstance(score.name, str) and score.name, f'Wrong score name "{score.name}"'
-    return score.name, score.value
-
-def convert_hyperparams_to_dict(hyperparams):
-    def _convert(p):
-        if p is None:
-            return None
-        d = {}
-        groups = getattr(p, 'groups', [])
-        parameters = getattr(p, 'parameters', [])
-        assert (not groups) or isinstance(groups, list), f'Wrong field "groups" of p={p}'
-        assert (not parameters) or isinstance(parameters, list), f'Wrong field "parameters" of p={p}'
-        for group_name in groups:
-            g = getattr(p, group_name, None)
-            d[group_name] = _convert(g)
-        for par_name in parameters:
-            d[par_name] = getattr(p, par_name, None)
-        return d
-    return _convert(hyperparams)
 
 def _create_object_detection_dataset_and_labels_schema(dataset_params):
     logger.debug(f'Using for train annotation file {dataset_params.annotations_train}')
@@ -348,999 +309,7 @@ def _create_object_detection_dataset_and_labels_schema(dataset_params):
     labels_schema = LabelSchemaEntity.from_labels(labels_list)
     return dataset, labels_schema
 
-class BaseOTETestAction(ABC):
-    _name = None
-    _with_validation = False
-    _depends_stages_names = []
 
-    @property
-    def name(self):
-        return type(self)._name
-
-    @property
-    def with_validation(self):
-        return type(self)._with_validation
-
-    @property
-    def depends_stages_names(self):
-        return type(self)._depends_stages_names
-
-    def __str__(self):
-        return (f'{type(self).__name__}('
-                f'name={self.name}, '
-                f'with_validation={self.with_validation}, '
-                f'depends_stages_names={self.depends_stages_names}'
-                f')')
-
-    def _check_result_prev_stages(self, results_prev_stages, list_required_stages):
-        for stage_name in list_required_stages:
-            if not results_prev_stages or stage_name not in results_prev_stages:
-                raise RuntimeError(f'The action {self.name} requires results of the stage {stage_name}, '
-                                   f'but they are absent')
-
-    @abstractmethod
-    def __call__(self, data_collector: DataCollector,
-                 results_prev_stages: Optional[OrderedDict]=None):
-        raise NotImplementedError('The main action method is not implemented')
-
-class OTETestTrainingAction(BaseOTETestAction):
-    _name = 'training'
-    def __init__(self, dataset, labels_schema, template_path, num_training_iters, batch_size):
-        self.dataset = dataset
-        self.labels_schema = labels_schema
-        self.template_path = template_path
-        self.num_training_iters = num_training_iters
-        self.batch_size = batch_size
-
-    @staticmethod
-    def _create_environment_and_task(params, labels_schema, model_template):
-        environment = TaskEnvironment(model=None, hyper_parameters=params, label_schema=labels_schema,
-                                      model_template=model_template)
-        logger.info('Create base Task')
-        task_impl_path = model_template.entrypoints.base
-        task_cls = get_task_class(task_impl_path)
-        task = task_cls(task_environment=environment)
-        return environment, task
-
-    def _get_training_performance_as_score_name_value(self):
-        training_performance = getattr(self.output_model, 'performance', None)
-        if training_performance is None:
-            raise RuntimeError('Cannot get training performance')
-        return performance_to_score_name_value(training_performance)
-
-    def _run_ote_training(self, data_collector):
-        logger.debug(f'self.template_path = {self.template_path}')
-
-        print(f'train dataset: {len(self.dataset.get_subset(Subset.TRAINING))} items')
-        print(f'validation dataset: {len(self.dataset.get_subset(Subset.VALIDATION))} items')
-
-        logger.debug('Load model template')
-        self.model_template = parse_model_template(self.template_path)
-
-        logger.debug('Set hyperparameters')
-        params = create(self.model_template.hyper_parameters.data)
-        if self.num_training_iters != KEEP_CONFIG_FIELD_VALUE():
-            params.learning_parameters.num_iters = int(self.num_training_iters)
-            logger.debug(f'Set params.learning_parameters.num_iters={params.learning_parameters.num_iters}')
-        else:
-            logger.debug(f'Keep params.learning_parameters.num_iters={params.learning_parameters.num_iters}')
-
-        if self.batch_size != KEEP_CONFIG_FIELD_VALUE():
-            params.learning_parameters.batch_size = int(self.batch_size)
-            logger.debug(f'Set params.learning_parameters.batch_size={params.learning_parameters.batch_size}')
-        else:
-            logger.debug(f'Keep params.learning_parameters.batch_size={params.learning_parameters.batch_size}')
-
-        logger.debug('Setup environment')
-        self.environment, self.task = self._create_environment_and_task(params,
-                                                                        self.labels_schema,
-                                                                        self.model_template)
-
-        logger.debug('Train model')
-        self.output_model = ModelEntity(
-            self.dataset,
-            self.environment.get_model_configuration(),
-            model_status=ModelStatus.NOT_READY)
-
-        self.copy_hyperparams = deepcopy(self.task._hyperparams)
-
-        self.task.train(self.dataset, self.output_model)
-        assert self.output_model.model_status == ModelStatus.SUCCESS, 'Training was failed'
-
-        score_name, score_value = self._get_training_performance_as_score_name_value()
-        logger.info(f'performance={self.output_model.performance}')
-        data_collector.log_final_metric('training_accuracy/' + score_name, score_value)
-
-        hyperparams_dict = convert_hyperparams_to_dict(self.copy_hyperparams)
-        for k, v in hyperparams_dict.items():
-            logger.debug(f'Try to run update_metadata:\nk={k}\nv={pformat(v, width=140)}\n')
-            data_collector.update_metadata(k, v)
-
-    def __call__(self, data_collector: DataCollector,
-                 results_prev_stages: Optional[OrderedDict]=None):
-        self._run_ote_training(data_collector)
-        results = {
-                'model_template': self.model_template,
-                'task': self.task,
-                'dataset': self.dataset,
-                'environment': self.environment,
-                'output_model': self.output_model,
-        }
-        return results
-
-def run_evaluation(dataset, task, model):
-    logger.debug('Evaluation: Get predictions on the dataset')
-    predicted_dataset = task.infer(
-        dataset.with_empty_annotations(),
-        InferenceParameters(is_evaluation=True))
-    resultset = ResultSetEntity(
-        model=model,
-        ground_truth_dataset=dataset,
-        prediction_dataset=predicted_dataset,
-    )
-    logger.debug('Evaluation: Estimate quality on dataset')
-    task.evaluate(resultset)
-    evaluation_performance = resultset.performance
-    logger.info(f'Evaluation: performance={evaluation_performance}')
-    score_name, score_value = performance_to_score_name_value(evaluation_performance)
-    return score_name, score_value
-
-class OTETestTrainingEvaluationAction(BaseOTETestAction):
-    _name = 'training_evaluation'
-    _with_validation = True
-    _depends_stages_names = ['training']
-
-    def __init__(self, subset=Subset.TESTING):
-        self.subset = subset
-
-    def _run_ote_evaluation(self, data_collector,
-                            dataset, task, trained_model):
-        logger.info('Begin evaluation of trained model')
-        validation_dataset = dataset.get_subset(self.subset)
-        score_name, score_value = run_evaluation(validation_dataset, task, trained_model)
-        data_collector.log_final_metric('evaluation_accuracy/' + score_name, score_value)
-        logger.info(f'End evaluation of trained model, results: {score_name}: {score_value}')
-        return score_name, score_value
-
-    def __call__(self, data_collector: DataCollector,
-                 results_prev_stages: Optional[OrderedDict]=None):
-        self._check_result_prev_stages(results_prev_stages, self.depends_stages_names)
-
-        kwargs = {
-                'dataset': results_prev_stages['training']['dataset'],
-                'task': results_prev_stages['training']['task'],
-                'trained_model': results_prev_stages['training']['output_model'],
-        }
-
-        score_name, score_value = self._run_ote_evaluation(data_collector, **kwargs)
-        results = {
-                'metrics': {
-                    'accuracy': {
-                        score_name: score_value
-                    }
-                }
-        }
-        return results
-
-def run_export(environment, dataset, task, action_name):
-    logger.debug(f'For action "{action_name}": Copy environment for evaluation exported model')
-
-    environment_for_export = deepcopy(environment)
-
-    logger.debug(f'For action "{action_name}": Create exported model')
-    exported_model = ModelEntity(
-        dataset,
-        environment_for_export.get_model_configuration(),
-        model_status=ModelStatus.NOT_READY)
-    logger.debug('Run export')
-    task.export(ExportType.OPENVINO, exported_model)
-
-    assert exported_model.model_status == ModelStatus.SUCCESS, \
-            f'In action "{action_name}": Export to OpenVINO was not successful'
-    assert exported_model.model_format == ModelFormat.OPENVINO, \
-            f'In action "{action_name}": Wrong model format after export'
-    assert exported_model.optimization_type == ModelOptimizationType.MO, \
-            f'In action "{action_name}": Wrong optimization type'
-
-    logger.debug(f'For action "{action_name}": Set exported model into environment for export')
-    environment_for_export.model = exported_model
-    return environment_for_export, exported_model
-
-class OTETestExportAction(BaseOTETestAction):
-    _name = 'export'
-    _depends_stages_names = ['training']
-
-    def _run_ote_export(self, data_collector,
-                        environment, dataset, task):
-        self.environment_for_export, self.exported_model = \
-                run_export(environment, dataset, task, action_name=self.name)
-
-    def __call__(self, data_collector: DataCollector,
-                 results_prev_stages: Optional[OrderedDict]=None):
-        self._check_result_prev_stages(results_prev_stages, self.depends_stages_names)
-
-        kwargs = {
-                'environment': results_prev_stages['training']['environment'],
-                'dataset': results_prev_stages['training']['dataset'],
-                'task': results_prev_stages['training']['task'],
-        }
-
-        self._run_ote_export(data_collector, **kwargs)
-        results = {
-                'environment': self.environment_for_export,
-                'exported_model': self.exported_model,
-        }
-        return results
-
-def create_openvino_task(model_template, environment):
-    logger.debug('Create OpenVINO Task')
-    openvino_task_impl_path = model_template.entrypoints.openvino
-    openvino_task_cls = get_task_class(openvino_task_impl_path)
-    openvino_task = openvino_task_cls(environment)
-    return openvino_task
-
-class OTETestExportEvaluationAction(BaseOTETestAction):
-    _name = 'export_evaluation'
-    _with_validation = True
-    _depends_stages_names = ['training', 'export', 'training_evaluation']
-
-    def __init__(self, subset=Subset.TESTING):
-        self.subset = subset
-
-    def _run_ote_export_evaluation(self, data_collector,
-                                   model_template, dataset,
-                                   environment_for_export, exported_model):
-        logger.info('Begin evaluation of exported model')
-        self.openvino_task = create_openvino_task(model_template, environment_for_export)
-        validation_dataset = dataset.get_subset(self.subset)
-        score_name, score_value = run_evaluation(validation_dataset, self.openvino_task, exported_model)
-        data_collector.log_final_metric('evaluation_accuracy_exported/' + score_name, score_value)
-        logger.info('End evaluation of exported model')
-        return score_name, score_value
-
-    def __call__(self, data_collector: DataCollector,
-                 results_prev_stages: Optional[OrderedDict]=None):
-        self._check_result_prev_stages(results_prev_stages, self.depends_stages_names)
-
-        kwargs = {
-                'model_template': results_prev_stages['training']['model_template'],
-                'dataset': results_prev_stages['training']['dataset'],
-                'environment_for_export': results_prev_stages['export']['environment'],
-                'exported_model': results_prev_stages['export']['exported_model'],
-        }
-
-        score_name, score_value = self._run_ote_export_evaluation(data_collector, **kwargs)
-        results = {
-                'metrics': {
-                    'accuracy': {
-                        score_name: score_value
-                    }
-                }
-        }
-        return results
-
-class OTETestPotAction(BaseOTETestAction):
-    _name = 'pot'
-    _depends_stages_names = ['export']
-
-    def __init__(self, pot_subset=Subset.TRAINING):
-        self.pot_subset = pot_subset
-
-    def _run_ote_pot(self, data_collector,
-                     model_template, dataset,
-                     environment_for_export):
-        logger.debug('Creating environment and task for POT optimization')
-        self.environment_for_pot = deepcopy(environment_for_export)
-        self.openvino_task_pot = create_openvino_task(model_template, environment_for_export)
-
-        self.optimized_model_pot = ModelEntity(
-            dataset,
-            self.environment_for_pot.get_model_configuration(),
-            model_status=ModelStatus.NOT_READY)
-        logger.info('Run POT optimization')
-        self.openvino_task_pot.optimize(
-            OptimizationType.POT,
-            dataset.get_subset(self.pot_subset),
-            self.optimized_model_pot,
-            OptimizationParameters())
-        assert self.optimized_model_pot.model_status == ModelStatus.SUCCESS, 'POT optimization was not successful'
-        assert self.optimized_model_pot.model_format == ModelFormat.OPENVINO, 'Wrong model format after pot'
-        assert self.optimized_model_pot.optimization_type == OptimizationType.POT, 'Wrong optimization type'
-        logger.info('POT optimization is finished')
-
-    def __call__(self, data_collector: DataCollector,
-                 results_prev_stages: Optional[OrderedDict]=None):
-        self._check_result_prev_stages(results_prev_stages, self.depends_stages_names)
-
-        kwargs = {
-                'model_template': results_prev_stages['training']['model_template'],
-                'dataset': results_prev_stages['training']['dataset'],
-                'environment_for_export': results_prev_stages['export']['environment'],
-        }
-
-        self._run_ote_pot(data_collector, **kwargs)
-        results = {
-                'openvino_task_pot': self.openvino_task_pot,
-                'optimized_model_pot': self.optimized_model_pot,
-        }
-        return results
-
-class OTETestPotEvaluationAction(BaseOTETestAction):
-    _name = 'pot_evaluation'
-    _with_validation = True
-    _depends_stages_names = ['training', 'pot', 'export_evaluation']
-
-    def __init__(self, subset=Subset.TESTING):
-        self.subset = subset
-
-    def _run_ote_pot_evaluation(self, data_collector,
-                                dataset,
-                                openvino_task_pot,
-                                optimized_model_pot):
-        logger.info('Begin evaluation of pot model')
-        validation_dataset_pot = dataset.get_subset(self.subset)
-        score_name, score_value = run_evaluation(validation_dataset_pot, openvino_task_pot, optimized_model_pot)
-        data_collector.log_final_metric('evaluation_accuracy_pot/' + score_name, score_value)
-        logger.info('End evaluation of pot model')
-        return score_name, score_value
-
-    def __call__(self, data_collector: DataCollector,
-                 results_prev_stages: Optional[OrderedDict]=None):
-        self._check_result_prev_stages(results_prev_stages, self.depends_stages_names)
-
-        kwargs = {
-                'dataset': results_prev_stages['training']['dataset'],
-                'openvino_task_pot': results_prev_stages['pot']['openvino_task_pot'],
-                'optimized_model_pot': results_prev_stages['pot']['optimized_model_pot'],
-        }
-
-        score_name, score_value = self._run_ote_pot_evaluation(data_collector, **kwargs)
-        results = {
-                'metrics': {
-                    'accuracy': {
-                        score_name: score_value
-                    }
-                }
-        }
-        return results
-
-class OTETestNNCFAction(BaseOTETestAction):
-    _name = 'nncf'
-    _depends_stages_names = ['training']
-
-    def _run_ote_nncf(self, data_collector,
-                      model_template, dataset, trained_model,
-                      environment):
-        logger.debug('Get predictions on the validation set for exported model')
-        self.environment_for_nncf = deepcopy(environment)
-
-        logger.info('Create NNCF Task')
-        nncf_task_class_impl_path = model_template.entrypoints.nncf
-        if not nncf_task_class_impl_path:
-            pytest.skip('NNCF is not enabled for this template')
-
-        if not is_nncf_enabled():
-            pytest.skip('NNCF is not installed')
-
-        logger.info('Creating NNCF task and structures')
-        self.nncf_model = ModelEntity(
-            dataset,
-            self.environment_for_nncf.get_model_configuration(),
-            model_status=ModelStatus.NOT_READY)
-        self.nncf_model.set_data('weights.pth', trained_model.get_data('weights.pth'))
-
-        self.environment_for_nncf.model = self.nncf_model
-
-        nncf_task_cls = get_task_class(nncf_task_class_impl_path)
-        self.nncf_task = nncf_task_cls(task_environment=self.environment_for_nncf)
-
-        logger.info('Run NNCF optimization')
-        self.nncf_task.optimize(OptimizationType.NNCF,
-                                dataset,
-                                self.nncf_model,
-                                OptimizationParameters())
-        assert self.nncf_model.model_status == ModelStatus.SUCCESS, 'NNCF optimization was not successful'
-        assert self.nncf_model.optimization_type == OptimizationType.NNCF, 'Wrong optimization type'
-        assert self.nncf_model.model_format == ModelFormat.BASE_FRAMEWORK, 'Wrong model format'
-        logger.info('NNCF optimization is finished')
-
-
-    def __call__(self, data_collector: DataCollector,
-                 results_prev_stages: Optional[OrderedDict]=None):
-        self._check_result_prev_stages(results_prev_stages, self.depends_stages_names)
-
-        kwargs = {
-                'model_template': results_prev_stages['training']['model_template'],
-                'dataset': results_prev_stages['training']['dataset'],
-                'trained_model': results_prev_stages['training']['output_model'],
-                'environment': results_prev_stages['training']['environment'],
-        }
-
-        self._run_ote_nncf(data_collector, **kwargs)
-        results = {
-                'nncf_task': self.nncf_task,
-                'nncf_model': self.nncf_model,
-                'nncf_environment': self.environment_for_nncf,
-        }
-        return results
-
-class OTETestNNCFEvaluationAction(BaseOTETestAction):
-    _name = 'nncf_evaluation'
-    _with_validation = True
-    _depends_stages_names = ['training', 'nncf', 'training_evaluation']
-
-    def __init__(self, subset=Subset.TESTING):
-        self.subset = subset
-
-    def _run_ote_nncf_evaluation(self, data_collector,
-                                dataset,
-                                nncf_task,
-                                nncf_model):
-        logger.info('Begin evaluation of nncf model')
-        validation_dataset = dataset.get_subset(self.subset)
-        score_name, score_value = run_evaluation(validation_dataset, nncf_task, nncf_model)
-        data_collector.log_final_metric('evaluation_accuracy_nncf/' + score_name, score_value)
-        logger.info('End evaluation of nncf model')
-        return score_name, score_value
-
-    def __call__(self, data_collector: DataCollector,
-                 results_prev_stages: Optional[OrderedDict]=None):
-        self._check_result_prev_stages(results_prev_stages, self.depends_stages_names)
-
-        kwargs = {
-                'dataset': results_prev_stages['training']['dataset'],
-                'nncf_task': results_prev_stages['nncf']['nncf_task'],
-                'nncf_model': results_prev_stages['nncf']['nncf_model'],
-        }
-
-        score_name, score_value = self._run_ote_nncf_evaluation(data_collector, **kwargs)
-        results = {
-                'metrics': {
-                    'accuracy': {
-                        score_name: score_value
-                    }
-                }
-        }
-        return results
-
-class OTETestNNCFExportAction(BaseOTETestAction):
-    _name = 'nncf_export'
-    _depends_stages_names = ['training', 'nncf']
-
-    def __init__(self, subset=Subset.VALIDATION):
-        self.subset = subset
-
-    def _run_ote_nncf_export(self, data_collector,
-                             nncf_environment, dataset, nncf_task):
-        logger.info('Begin export of nncf model')
-        self.environment_nncf_export, self.nncf_exported_model = \
-                run_export(nncf_environment, dataset, nncf_task, action_name=self.name)
-        logger.info('End export of nncf model')
-
-    def __call__(self, data_collector: DataCollector,
-                 results_prev_stages: Optional[OrderedDict]=None):
-        self._check_result_prev_stages(results_prev_stages, self.depends_stages_names)
-
-        kwargs = {
-                'nncf_environment': results_prev_stages['nncf']['nncf_environment'],
-                'dataset': results_prev_stages['training']['dataset'],
-                'nncf_task': results_prev_stages['nncf']['nncf_task'],
-        }
-
-        self._run_ote_nncf_export(data_collector, **kwargs)
-        results = {
-                'environment': self.environment_nncf_export,
-                'exported_model': self.nncf_exported_model,
-        }
-        return results
-
-class OTETestNNCFExportEvaluationAction(BaseOTETestAction):
-    _name = 'nncf_export_evaluation'
-    _with_validation = True
-    _depends_stages_names = ['training', 'nncf_export', 'nncf_evaluation']
-
-    def __init__(self, subset=Subset.TESTING):
-        self.subset = subset
-
-    def _run_ote_nncf_export_evaluation(self, data_collector,
-                                        model_template, dataset,
-                                        nncf_environment_for_export, nncf_exported_model):
-        logger.info('Begin evaluation of NNCF exported model')
-        self.openvino_task = create_openvino_task(model_template, nncf_environment_for_export)
-        validation_dataset = dataset.get_subset(self.subset)
-        score_name, score_value = run_evaluation(validation_dataset, self.openvino_task, nncf_exported_model)
-        data_collector.log_final_metric('evaluation_accuracy_nncf_exported/' + score_name, score_value)
-        logger.info('End evaluation of NNCF exported model')
-        return score_name, score_value
-
-    def __call__(self, data_collector: DataCollector,
-                 results_prev_stages: Optional[OrderedDict]=None):
-        self._check_result_prev_stages(results_prev_stages, self.depends_stages_names)
-
-        kwargs = {
-                'model_template': results_prev_stages['training']['model_template'],
-                'dataset': results_prev_stages['training']['dataset'],
-                'nncf_environment_for_export': results_prev_stages['nncf_export']['environment'],
-                'nncf_exported_model': results_prev_stages['nncf_export']['exported_model'],
-        }
-
-        score_name, score_value = self._run_ote_nncf_export_evaluation(data_collector, **kwargs)
-        results = {
-                'metrics': {
-                    'accuracy': {
-                        score_name: score_value
-                    }
-                }
-        }
-        return results
-
-def get_value_from_dict_by_dot_separated_address(struct, address):
-    def _get(cur_struct, addr):
-        assert isinstance(addr, list)
-        if not addr:
-            return cur_struct
-        assert isinstance(cur_struct, dict)
-        if addr[0] not in cur_struct:
-            raise ValueError(f'Cannot find address {address} in struct {struct}: {addr[0]} is absent in {cur_struct}')
-        return _get(cur_struct[addr[0]], addr[1:])
-
-    assert isinstance(address, str), f'The parameter address should be string, address={address}'
-    return _get(struct, address.split('.'))
-
-class Validator:
-    """
-    The class receives info on results metric of the current test stage and
-    compares it with the expected metrics.
-    """
-    def __init__(self, cur_test_expected_metrics_callback: Union[None, Callable[[],Dict]]):
-        self.cur_test_expected_metrics_callback = cur_test_expected_metrics_callback
-
-    # TODO(lbeynens): add a method to extract dependency info from expected metrics
-    #                 to add the stages we depend on to the dependency list.
-
-    @staticmethod
-    def _get_min_max_value_from_expected_metrics(cur_metric_requirements: Dict,
-                                                 test_results_storage: Dict):
-        """
-        The method gets requirement for some metric and convert it to the triplet
-        (target_value, min_value, max_value).
-        Note that the target_value may be pointed either by key 'target_value' (in this case it is float),
-        or by the key 'base', in this case it is a dot-separated address to another value in the
-        storage of previous stages' results `test_results_storage`.
-
-        Note that the range for the metric values may be pointed by key 'max_diff',
-        in this case the range will be [target_value - max_diff, target_value + max_diff]
-        (inclusively).
-
-        But also the range may be pointed by keys 'max_diff_if_less_threshold' and
-        'max_diff_if_greater_threshold', in this case the range is
-        [target_value - max_diff_if_less_threshold, target_value + max_diff_if_greater_threshold]
-        (also inclusively). This allows to point non-symmetric ranges w.r.t. the target_value.
-
-        Also note that if one of 'max_diff_if_less_threshold' and 'max_diff_if_greater_threshold'
-        is absent, it is set to `+infinity`, so the range will be bounded from one side
-        (but not both of them, this will be an error)
-        """
-        keys = set(cur_metric_requirements.keys())
-        if 'target_value' not in keys and 'base' not in keys:
-            raise ValueError(f'Wrong cur_metric_requirements: either "target_value" or "base" '
-                             f' should be pointed in the structure, whereas '
-                             f'cur_metric_requirements={pformat(cur_metric_requirements)}')
-        if 'target_value' in keys and 'base' in keys:
-            raise ValueError(f'Wrong cur_metric_requirements: either "target_value" or "base" '
-                             f' should be pointed in the structure, but not both, whereas '
-                             f'cur_metric_requirements={pformat(cur_metric_requirements)}')
-        if ('max_diff' not in keys) and ('max_diff_if_less_threshold' not in keys) \
-                and ('max_diff_if_greater_threshold' not in keys):
-            raise ValueError(f'Wrong cur_metric_requirements: either "max_diff" or one/two of '
-                             f'"max_diff_if_less_threshold" and "max_diff_if_greater_threshold" should be '
-                             f'pointed in the structure, whereas '
-                             f'cur_metric_requirements={pformat(cur_metric_requirements)}')
-
-        if ('max_diff' in keys) and ('max_diff_if_less_threshold' in keys or 'max_diff_if_greater_threshold' in keys):
-            raise ValueError(f'Wrong cur_metric_requirements: either "max_diff" or one/two of '
-                             f'"max_diff_if_less_threshold" and "max_diff_if_greater_threshold" should be '
-                             f'pointed in the structure, but not both, whereas '
-                             f'cur_metric_requirements={pformat(cur_metric_requirements)}')
-
-        if 'target_value' in cur_metric_requirements:
-            target_value = float(cur_metric_requirements['target_value'])
-        elif 'base' in cur_metric_requirements:
-            base_metric_address = cur_metric_requirements['base']
-            target_value = get_value_from_dict_by_dot_separated_address(test_results_storage, base_metric_address)
-            target_value = float(target_value)
-        else:
-            raise RuntimeError(f'ERROR: Wrong parsing of metric requirements {cur_metric_requirements}')
-
-        if 'max_diff' in cur_metric_requirements:
-            max_diff = cur_metric_requirements['max_diff']
-            max_diff = float(max_diff)
-            if not max_diff >= 0:
-                raise ValueError(f'Wrong max_diff {max_diff} -- it should be a non-negative number')
-            return (target_value, target_value - max_diff, target_value + max_diff)
-
-        max_diff_if_less_threshold = cur_metric_requirements.get('max_diff_if_less_threshold')
-        max_diff_if_greater_threshold = cur_metric_requirements.get('max_diff_if_greater_threshold')
-        if max_diff_if_less_threshold is None and max_diff_if_greater_threshold is None:
-            raise ValueError(f'Wrong cur_metric_requirements: all of max_diff, max_diff_if_less_threshold, and '
-                             f'max_diff_if_greater_threshold are None, '
-                             f'cur_metric_requirements={pformat(cur_metric_requirements)}')
-
-        if max_diff_if_greater_threshold is not None:
-            max_diff_if_greater_threshold = float(max_diff_if_greater_threshold)
-            if not max_diff_if_greater_threshold >= 0:
-                raise ValueError(f'Wrong max_diff_if_greater_threshold {max_diff_if_greater_threshold} '
-                                 f'-- it should be a non-negative number')
-
-            max_value = target_value + max_diff_if_greater_threshold
-        else:
-            max_value = None
-
-        if max_diff_if_less_threshold is not None:
-            max_diff_if_less_threshold = float(max_diff_if_less_threshold)
-            if not max_diff_if_less_threshold >= 0:
-                raise ValueError(f'Wrong max_diff_if_less_threshold {max_diff_if_less_threshold} '
-                                 f'-- it should be a non-negative number')
-
-            min_value = target_value - max_diff_if_less_threshold
-        else:
-            min_value = None
-
-        return (target_value, min_value, max_value)
-
-    @staticmethod
-    def _compare(current_metric: float, cur_res_addr: str,
-                 target_value: float, min_value: Union[float, None], max_value: Union[float, None]):
-        assert all(isinstance(v, float) for v in [current_metric, target_value] )
-        assert all((v is None) or isinstance(v, float) for v in [min_value, max_value])
-
-        if min_value is not None and max_value is not None:
-            assert min_value <= target_value <= max_value
-
-            if min_value <= current_metric <= max_value:
-                logger.info(f'Validation: passed: The metric {cur_res_addr} is in the acceptable range '
-                            f'near the target value {target_value}: '
-                            f'{current_metric} is in [{min_value}, {max_value}]')
-                is_passed = True
-                cur_fail_reason = None
-            else:
-                cur_fail_reason = (f'Validation: failed: The metric {cur_res_addr} is NOT in the acceptable range '
-                                   f'near the target value {target_value}: '
-                                   f'{current_metric} is NOT in [{min_value}, {max_value}]')
-                logger.error(cur_fail_reason)
-                is_passed = False
-            return is_passed, cur_fail_reason
-
-        assert (min_value is not None) or (max_value is not None)
-        if min_value is not None:
-            cmp_op = lambda x: x >= min_value
-            cmp_str_true = 'greater or equal'
-            cmp_op_str_true = '>='
-            cmp_op_str_false = '<'
-            threshold = min_value
-        else:
-            cmp_op = lambda x: x <= max_value
-            cmp_str_true = 'less or equal'
-            cmp_op_str_true = '<='
-            cmp_op_str_false = '>'
-            threshold = max_value
-        acceptable_error = abs(threshold - target_value)
-        if cmp_op(current_metric):
-            logger.info(f'Validation: passed: The metric {cur_res_addr} is {cmp_str_true} '
-                        f'the target value {target_value} with acceptable error {acceptable_error}: '
-                        f'{current_metric} {cmp_op_str_true} {threshold}')
-            is_passed = True
-            cur_fail_reason = None
-        else:
-            cur_fail_reason = (f'Validation: failed: The metric {cur_res_addr} is NOT {cmp_str_true} '
-                               f'the target value {target_value} with acceptable error {acceptable_error}: '
-                               f'{current_metric} {cmp_op_str_false} {threshold}')
-            logger.error(cur_fail_reason)
-            is_passed = False
-        return is_passed, cur_fail_reason
-
-    def validate(self, current_result: Dict, test_results_storage: Dict):
-        """
-        The method validates results of the current test.
-        :param current_result -- dict with result of the current test
-        :param test_results_storage -- dict with results of previous tests
-                                       of this test case
-                                       (e.g. the same training parameters)
-
-        The function returns nothing, but may raise exceptions to fail the test.
-        If the structure stored expected metrics is wrong, the function raises ValueError.
-        """
-        if self.cur_test_expected_metrics_callback is None:
-            # most probably, it is not a reallife test
-            logger.info(f'Validation: skipped, since there should not be expected metrics for this test, '
-                        f'most probably the test is not run in "{REALLIFE_USECASE_CONSTANT()}" usecase')
-            return
-
-        logger.info('Validation: begin')
-
-        # calling the callback to receive expected metrics for the current test
-        cur_test_expected_metrics = self.cur_test_expected_metrics_callback()
-
-        assert isinstance(cur_test_expected_metrics, dict), \
-                f'Wrong current test expected metric: {cur_test_expected_metrics}'
-        logger.debug(f'Validation: received cur_test_expected_metrics={pformat(cur_test_expected_metrics)}')
-        is_passed = True
-        fail_reasons = []
-        for k, v in cur_test_expected_metrics.items():
-            # TODO(lbeynens): add possibility to point a list of requirements for a metric
-            cur_res_addr = k
-            cur_metric_requirements = v
-            logger.info(f'Validation: begin check {cur_res_addr}')
-            try:
-                current_metric = get_value_from_dict_by_dot_separated_address(current_result, cur_res_addr)
-                current_metric = float(current_metric)
-            except (ValueError, TypeError) as e:
-                raise ValueError(f'Cannot get metric {cur_res_addr} from the current result {current_result}') from e
-
-            logger.debug(f'current_metric = {current_metric}')
-            try:
-                target_value, min_value, max_value = \
-                        self._get_min_max_value_from_expected_metrics(cur_metric_requirements,
-                                                                      test_results_storage)
-            except (ValueError, TypeError) as e:
-                raise ValueError(f'Error when parsing expected metrics for the metric {cur_res_addr}') from e
-
-            cur_is_passed, cur_fail_reason = self._compare(current_metric, cur_res_addr,
-                                                           target_value, min_value, max_value)
-            if not cur_is_passed:
-                is_passed = False
-                fail_reasons.append(cur_fail_reason)
-
-            logger.info(f'Validation: end check {cur_res_addr}')
-
-        logger.info(f'Validation: end, result={is_passed}')
-        if not is_passed:
-            fail_reasons = '\n'.join(fail_reasons)
-            pytest.fail(f'Validation failed:\n{fail_reasons}')
-
-class OTETestStagesStorageInterface(ABC):
-    @abstractmethod
-    def get_stage(self, name: str) -> 'OTETestStage':
-        raise NotImplementedError('The method get_stage is not implemented')
-
-class OTETestStage:
-    """
-    OTETestStage -- auxiliary class that
-    1. Allows to set up dependency between test stages: before the main action of a test stage is run, all the actions
-       for the stages that are pointed in 'depends' list are called beforehand;
-    2. Runs for each test stage its main action only once: the main action is run inside try-except clause, and
-       2.1. if the action was executed without exceptions, a flag `was_processed` is set, the results of the action
-            are kept, and the next time the stage is called no action is executed;
-       2.2. if the action raised an exception, the exception is stored, the flag `was_processed` is set, and the next
-            time the stage is called the exception is re-raised.
-    """
-    def __init__(self, action: BaseOTETestAction,
-                 stages_storage: OTETestStagesStorageInterface):
-        self.was_processed = False
-        self.stored_exception = None
-        self.action = action
-        self.stages_storage = stages_storage
-        self.stage_results = {}
-        assert isinstance(self.stages_storage, OTETestStagesStorageInterface)
-        assert isinstance(self.action, BaseOTETestAction)
-
-    def __str__(self):
-        return (f'{type(self).__name__}('
-                f'action={self.action}, '
-                f'was_processed={self.was_processed}, '
-                f'stored_exception={self.stored_exception}, '
-                f'stage_results.keys={list(self.stage_results.keys())}, '
-                f'id(stages_storage)={id(self.stages_storage)}'
-                f')')
-
-    @property
-    def name(self):
-        return self.action.name
-
-    def get_depends_stages(self):
-        logger.debug(f'get_depends_stages for stage {self.name}: begin')
-        depends_stages_names = self.action.depends_stages_names
-        assert isinstance(depends_stages_names, list)
-        assert all(isinstance(v, str) for v in depends_stages_names)
-
-        stages = []
-        for stage_name in depends_stages_names:
-            logger.debug(f'get_depends_stages: get stage with name "{stage_name}"')
-            cur_stage = self.stages_storage.get_stage(stage_name)
-            assert isinstance(cur_stage, OTETestStage), f'Wrong stage for stage_name="{stage_name}"'
-            assert cur_stage.name == stage_name, \
-                    f'For stage_name="{stage_name}" got the stage with name="{cur_stage.name}"'
-            logger.debug(f'get_depends_stages: cur_stage="{cur_stage}"')
-            stages.append(cur_stage)
-        logger.debug(f'get_depends_stages for stage {self.name}: end')
-        return stages
-
-    def _reraise_stage_exception_if_was_failed(self):
-        assert self.was_processed, \
-                'The method _reraise_stage_exception_if_was_failed should be used only for stages that were processed'
-        if self.stored_exception is None:
-            # nothing to do here
-            return
-
-        logger.warning(f'In stage {self.name}: found that previous call of the stage '
-                       'caused exception -- re-raising it')
-        raise self.stored_exception
-
-    def _run_validation(self, test_results_storage: Dict,
-                        validator: Union[Validator, None]):
-        if not self.action.with_validation:
-            return
-        if validator is None:
-            logger.debug('The validator is None -- the validation should be skipped, '
-                         'most probably this test stage was run from a dependency chain')
-            return
-
-        validator.validate(self.stage_results, test_results_storage)
-
-    def run_once(self, data_collector: DataCollector, test_results_storage: OrderedDict,
-                 validator: Union[Validator, None]):
-        logger.info(f'Begin stage "{self.name}"')
-        assert isinstance(test_results_storage, OrderedDict)
-        logger.debug(f'For test stage "{self.name}": test_results_storage.keys = {list(test_results_storage.keys())}')
-
-        for dep_stage in self.get_depends_stages():
-            # Processing all dependency stages of the current test.
-            # Note that
-            # * the stages may run their own dependency stages -- they will compose so called "dependency chain"
-            # * the dependency stages are run with `validator = None`
-            #   to avoid validation of stages that are run from the dependency chain.
-            logger.debug(f'For test stage "{self.name}": Before running dep. stage "{dep_stage.name}"')
-            dep_stage.run_once(data_collector, test_results_storage, validator=None)
-            logger.debug(f'For test stage "{self.name}": After running dep. stage "{dep_stage.name}"')
-
-        if self.was_processed:
-            self._reraise_stage_exception_if_was_failed()
-            # if we are here, then the stage was processed without exceptions
-            logger.info(f'The stage {self.name} was already processed SUCCESSFULLY')
-
-            # Run validation here for the rare case if this test now is being run *not* from a dependency chain
-            # (i.e. the test is run with `validator != None`),
-            # but the test already has been run from some dependency chain earlier.
-            self._run_validation(test_results_storage, validator)
-
-            logger.info(f'End stage "{self.name}"')
-            return
-
-        if self.name in test_results_storage:
-            raise RuntimeError(f'Error: For test stage "{self.name}": '
-                               f'another OTETestStage with name {self.name} has been run already')
-
-        try:
-            logger.info(f'For test stage "{self.name}": Before running main action')
-            self.stage_results = self.action(data_collector=data_collector,
-                                             results_prev_stages=test_results_storage)
-            logger.info(f'For test stage "{self.name}": After running main action')
-            self.was_processed = True
-            test_results_storage[self.name] = self.stage_results
-            logger.debug(f'For test stage "{self.name}": after addition test_results_storage.keys = '
-                         f'{list(test_results_storage.keys())}')
-        except Exception as e:
-            logger.info(f'For test stage "{self.name}": After running action for stage {self.name} -- CAUGHT EXCEPTION:\n{e}')
-            logger.info(f'End stage "{self.name}"')
-            self.stored_exception = e
-            self.was_processed = True
-            raise e
-
-        # The validation step is made outside the central try...except clause, since if the test was successful, but
-        # the quality numbers were lower than expected, the result of the stage still may be re-used
-        # in other stages.
-        self._run_validation(test_results_storage, validator)
-        logger.info(f'End stage "{self.name}"')
-
-def _get_duplications(arr):
-    c = Counter(arr)
-    dups = [k for k, v in c.items() if v > 1]
-    return dups
-
-def _str_dict_with_shortened_vals(d, max_len=200):
-    assert isinstance(d, dict)
-    if not d:
-        return '{}'
-    def _shorten(v):
-        sv = str(v)
-        if len(sv) <= max_len:
-            return sv
-        return sv[:max_len] + '...'
-
-    s = '\n'.join(f'{k}: {_shorten(v)},' for k, v in d.items())
-    s = '\n    '.join(s.split('\n'))
-    s = '{\n    ' + s + '\n}'
-    return s
-
-class OTETestCaseInterface(OTETestStagesStorageInterface):
-    @classmethod
-    @abstractmethod
-    def get_list_of_test_stages(cls):
-        raise NotImplementedError('The method get_list_of_test_stages is not implemented')
-
-    @abstractmethod
-    def run_stage(self, stage_name: str, data_collector: DataCollector,
-                  validator: Validator):
-        raise NotImplementedError('The method run_stage is not implemented')
-
-def generate_ote_integration_test_case_class(test_actions_classes: List[Type[BaseOTETestAction]]) -> Type:
-    test_actions_classes = deepcopy(test_actions_classes)
-
-    # check names' duplication
-    classes_names = [action_cls._name for action_cls in test_actions_classes]
-    name_dups = _get_duplications(classes_names)
-    if name_dups:
-        raise ValueError(f'Wrong input: there are duplications in names of actions; duplications = {name_dups}')
-
-    class _OTEIntegrationTestCase(OTETestCaseInterface):
-        _TEST_STAGES = [action_cls._name for action_cls in test_actions_classes]
-
-        @classmethod
-        def get_list_of_test_stages(cls):
-            return deepcopy(cls._TEST_STAGES)
-
-        def __init__(self, params_factories_for_test_actions: Dict[str, Callable[[], Dict]]):
-            logger.debug(f'initialization of test case: begin')
-            self._stages = OrderedDict()
-            for action_cls in test_actions_classes:
-                logger.debug(f'initialization of test case: action_cls={action_cls}')
-
-                cur_name = action_cls._name
-                cur_params_factory = params_factories_for_test_actions.get(cur_name)
-                if cur_params_factory is not None:
-                    logger.debug(f'initialization of test case: calling params factory')
-                    cur_params = cur_params_factory()
-                else:
-                    cur_params = {}
-
-                assert isinstance(cur_params, dict), f'Wrong params received from factory: {cur_params}'
-                short_params_str = _str_dict_with_shortened_vals(cur_params)
-                logger.info(f'initialization of test case: add action "{cur_name}" with params={short_params_str}')
-
-                cur_action = action_cls(**cur_params)
-
-                # Note that `self` is used as stages_storage for OTETestStage below
-                cur_stage = OTETestStage(action=cur_action,
-                                         stages_storage=self)
-                self._stages[cur_name] = cur_stage
-
-            assert list(self._stages.keys()) == list(self._TEST_STAGES)
-
-            # test results should be kept between stages
-            self.test_results_storage = OrderedDict()
-            logger.debug(f'initialization of test case: end')
-
-        # implementation of method from OTETestStagesStorageInterface
-        def get_stage(self, name: str) -> 'OTETestStage':
-            return self._stages[name]
-
-        def run_stage(self, stage_name: str, data_collector: DataCollector,
-                      validator: Validator):
-            assert stage_name in self._TEST_STAGES, f'Wrong stage_name {stage_name}'
-            self._stages[stage_name].run_once(data_collector, self.test_results_storage,
-                                              validator)
-
-    return _OTEIntegrationTestCase
-
-def get_default_test_action_classes() -> List[Type[BaseOTETestAction]]:
-    return [OTETestTrainingAction,
-            OTETestTrainingEvaluationAction,
-            OTETestExportAction,
-            OTETestExportEvaluationAction,
-            OTETestPotAction,
-            OTETestPotEvaluationAction,
-            OTETestNNCFAction,
-            OTETestNNCFEvaluationAction,
-            OTETestNNCFExportAction,
-            OTETestNNCFExportEvaluationAction,
-    ]
-
-
-class OTETrainingTestInterface(ABC):
-    @classmethod
-    @abstractmethod
-    def get_list_of_tests(cls, usecase: Optional[str] = None):
-        raise NotImplementedError('The method is not implemented')
 
 
 # pytest magic
@@ -1355,188 +324,6 @@ def pytest_generate_tests(metafunc):
 
     argnames, argvalues, ids = metafunc.cls.get_list_of_tests(usecase)
     metafunc.parametrize(argnames, argvalues, ids=ids, scope='class')
-
-class OTETestCreationParametersInterface(ABC):
-    @abstractmethod
-    def test_bunches(self) -> List[Dict[str, Any]]:
-        raise NotImplementedError('The method is not implemented')
-
-    @abstractmethod
-    def test_case_class(self) -> Type[OTETestCaseInterface]:
-        raise NotImplementedError('The method is not implemented')
-
-    @abstractmethod
-    def short_test_parameters_names_for_generating_id(self) -> OrderedDict:
-        raise NotImplementedError('The method is not implemented')
-
-    @abstractmethod
-    def test_parameters_defining_impl_behavior(self) -> List[str]:
-        raise NotImplementedError('The method is not implemented')
-
-    @abstractmethod
-    def default_test_parameters(self) -> Dict[str, Any]:
-        raise NotImplementedError('The method is not implemented')
-
-class DefaultOTETestCreationParametersInterface(OTETestCreationParametersInterface):
-    def test_case_class(self) -> Type[OTETestCaseInterface]:
-        return generate_ote_integration_test_case_class(get_default_test_action_classes())
-
-    def short_test_parameters_names_for_generating_id(self) -> OrderedDict:
-        DEFAULT_SHORT_TEST_PARAMETERS_NAMES_FOR_GENERATING_ID = OrderedDict([
-                ('test_stage', 'ACTION'),
-                ('model_name', 'model'),
-                ('dataset_name', 'dataset'),
-                ('num_training_iters', 'num_iters'),
-                ('batch_size', 'batch'),
-                ('usecase', 'usecase'),
-        ])
-        return deepcopy(DEFAULT_SHORT_TEST_PARAMETERS_NAMES_FOR_GENERATING_ID)
-
-    def test_parameters_defining_impl_behavior(self) -> List[str]:
-        DEFAULT_TEST_PARAMETERS_DEFINING_IMPL_BEHAVIOR = ['model_name',
-                                                          'dataset_name',
-                                                          'num_training_iters',
-                                                          'batch_size']
-        return deepcopy(DEFAULT_TEST_PARAMETERS_DEFINING_IMPL_BEHAVIOR)
-
-    def default_test_parameters(self) -> Dict[str, Any]:
-        DEFAULT_TEST_PARAMETERS = {
-                'num_iters': 1,
-                'batch_size': 2,
-        }
-        return deepcopy(DEFAULT_TEST_PARAMETERS)
-
-class OTETestHelper:
-    class _Cache:
-        def __init__(self):
-            self._cache_parameters = {}
-            self._cached_value = None
-        def get(self):
-            return self._cached_value
-        def set(self, params, value):
-            logger.debug(f'cache.set new value for parameters {params}')
-            self._cache_parameters = deepcopy(params)
-            self._cached_value = value
-        def has_same_params(self, params):
-            res = (self._cache_parameters == params)
-            res_str = '==' if res else '!='
-            logger.debug(f'cache.has_same_params: '
-                         f'cache_parameters={self._cache_parameters} {res_str} {params}, res={res}')
-            return res
-
-    def __init__(self,
-                 test_creation_parameters):
-        assert isinstance(test_creation_parameters, OTETestCreationParametersInterface)
-
-        self.test_case_class = test_creation_parameters.test_case_class()
-        self.test_bunches = test_creation_parameters.test_bunches()
-
-        self.short_test_parameters_names_for_generating_id = \
-                test_creation_parameters.short_test_parameters_names_for_generating_id()
-        #TODO(beynens): rename to test_parameters_defining_test_case
-        self.test_parameters_defining_impl_behavior = \
-                test_creation_parameters.test_parameters_defining_impl_behavior()
-        self.default_test_parameters = \
-                test_creation_parameters.default_test_parameters()
-
-        self._cache = OTETestHelper._Cache()
-
-        assert issubclass(self.test_case_class, OTETestCaseInterface)
-        assert isinstance(self.short_test_parameters_names_for_generating_id, OrderedDict)
-        assert all(isinstance(k, str) and isinstance(v, str)
-                   for k, v in self.short_test_parameters_names_for_generating_id.items())
-        assert 'test_stage' in self.short_test_parameters_names_for_generating_id
-        assert 'model_name' in self.short_test_parameters_names_for_generating_id
-        assert 'dataset_name' in self.short_test_parameters_names_for_generating_id
-        assert 'usecase' in self.short_test_parameters_names_for_generating_id
-
-        assert isinstance(self.test_parameters_defining_impl_behavior, list)
-        assert all(isinstance(s, str) for s in self.test_parameters_defining_impl_behavior)
-
-        assert isinstance(self.default_test_parameters, dict)
-        assert all(isinstance(k, str) for k in self.default_test_parameters.keys())
-
-    def _get_list_of_test_stages(self):
-        return self.test_case_class.get_list_of_test_stages()
-
-    def _fill_test_parameters_default_values(self, test_parameters):
-        for key, default_val in self.default_test_parameters.items():
-            val = test_parameters.get(key)
-            if val is None or val == DEFAULT_FIELD_VALUE_FOR_USING_IN_TEST():
-                test_parameters[key] = default_val
-
-    def _generate_test_id(self, test_parameters):
-        id_parts = (
-                f'{short_par_name}-{test_parameters[par_name]}'
-                for par_name, short_par_name in self.short_test_parameters_names_for_generating_id.items()
-        )
-        return ','.join(id_parts)
-
-    def get_list_of_tests(self, usecase: Optional[str] = None):
-        """
-        The functions generates the lists of values for the tests from the field test_bunches of the class.
-
-        The function returns two lists
-        * argnames -- a tuple with names of the test parameters, at the moment it is
-                      a one-element tuple with the parameter name "test_parameters"
-        * argvalues -- list of tuples, each tuple has the same len as argname tuple,
-                       at the moment it is a one-element tuple with the dict `test_parameters`
-                       that stores the parameters of the test
-        * ids -- list of strings with ids corresponding the parameters of the tests
-                 each id is a string generated from the corresponding test_parameters
-                 value -- see the functions _generate_test_id
-
-        The lists argvalues and ids will have the same length.
-
-        If the parameter `usecase` is set, it makes filtering by usecase field of test bunches.
-        """
-        test_bunches = self.test_bunches
-        assert all(isinstance(el, dict) for el in test_bunches)
-
-        argnames = ('test_parameters',)
-        argvalues = []
-        ids = []
-        for el in test_bunches:
-            el_model_name = el.get('model_name')
-            el_dataset_name = el.get('dataset_name')
-            el_usecase = el.get('usecase')
-            if usecase is not None and el_usecase != usecase:
-                continue
-            if isinstance(el_model_name, (list, tuple)):
-                model_names = el_model_name
-            else:
-                model_names = [el_model_name]
-            if isinstance(el_dataset_name, (list, tuple)):
-                dataset_names = el_dataset_name
-            else:
-                dataset_names = [el_dataset_name]
-
-            model_dataset_pairs = list(itertools.product(model_names, dataset_names))
-
-            for m, d in model_dataset_pairs:
-                for test_stage in self._get_list_of_test_stages():
-                    test_parameters = deepcopy(el)
-                    test_parameters['test_stage'] = test_stage
-                    test_parameters['model_name'] = m
-                    test_parameters['dataset_name'] = d
-                    self._fill_test_parameters_default_values(test_parameters)
-                    argvalues.append((test_parameters,))
-                    ids.append(self._generate_test_id(test_parameters))
-
-        return argnames, argvalues, ids
-
-    def get_test_case(self, test_parameters, params_factories_for_test_actions):
-        params_defining_cache = {k: test_parameters[k] for k in self.test_parameters_defining_impl_behavior}
-        if not self._cache.has_same_params(params_defining_cache):
-            logger.info(f'OTETestHelper: parameters were changed -- updating cache')
-            logger.info(f'OTETestHelper: before creating test case (class {self.test_case_class})')
-            test_case = self.test_case_class(params_factories_for_test_actions)
-            logger.info(f'OTETestHelper: after creating test case (class {self.test_case_class})')
-            self._cache.set(params_defining_cache, test_case)
-        else:
-            logger.info('OTETestHelper: parameters were not changed -- cache is kept')
-
-        return self._cache.get()
 
 
 class ObjectDetectionTrainingTestParameters(DefaultOTETestCreationParametersInterface):
@@ -1655,6 +442,5 @@ class TestOTEReallifeObjectDetection(OTETrainingTestInterface):
              test_parameters,
              test_case_fx, data_collector_fx,
              cur_test_expected_metrics_callback_fx):
-        validator = Validator(cur_test_expected_metrics_callback_fx)
         test_case_fx.run_stage(test_parameters['test_stage'], data_collector_fx,
-                               validator)
+                               cur_test_expected_metrics_callback_fx)
